// pronti-frontend-main/front/src/app/modulos/cuenta-contable/form/form-cuenta-contable.enum.ts
export enum FormCuentaContableEnum {
  idPeriodoContable = 'idPeriodoContable',
  idRubrosRol = 'idRubrosRol',
  idAgencia = 'idAgencia',
  tipo = 'tipo',
  nombre = 'nombre',
  identificador = 'identificador',
  nivel = 'nivel',
  idNivel1 = 'idNivel1',
  idNivel2 = 'idNivel2',
  idNivel3 = 'idNivel3',
  idNivel4 = 'idNivel4',
  idNivel5 = 'idNivel5',
  tipoCuenta = 'tipoCuenta',
  movimiento = 'movimiento',
  anteriorDebito = 'anteriorDebito',
  anteriorCredito = 'anteriorCredito',
  anteriorSaldo = 'anteriorSaldo',
  eneroDebito = 'eneroDebito',
  eneroCredito = 'eneroCredito',
  eneroSaldo = 'eneroSaldo',
  febreroDebito = 'febreroDebito',
  febreroCredito = 'febreroCredito',
  febreroSaldo = 'febreroSaldo',
  marzoDebito = 'marzoDebito',
  marzoCredito = 'marzoCredito',
  marzoSaldo = 'marzoSaldo',
  abrilDebito = 'abrilDebito',
  abrilCredito = 'abrilCredito',
  abrilSaldo = 'abrilSaldo',
  mayoDebito = 'mayoDebito',
  mayoCredito = 'mayoCredito',
  mayoSaldo = 'mayoSaldo',
  junioDebito = 'junioDebito',
  junioCredito = 'junioCredito',
  junioSaldo = 'junioSaldo',
  julioDebito = 'julioDebito',
  julioCredito = 'julioCredito',
  julioSaldo = 'julioSaldo',
  agostoDebito = 'agostoDebito',
  agostoCredito = 'agostoCredito',
  agostoSaldo = 'agostoSaldo',
  septiembreDebito = 'septiembreDebito',
  septiembreCredito = 'septiembreCredito',
  septiembreSaldo = 'septiembreSaldo',
  octubreDebito = 'octubreDebito',
  octubreCredito = 'octubreCredito',
  octubreSaldo = 'octubreSaldo',
  noviembreDebito = 'noviembreDebito',
  noviembreCredito = 'noviembreCredito',
  noviembreSaldo = 'noviembreSaldo',
  diciembreDebito = 'diciembreDebito',
  diciembreCredito = 'diciembreCredito',
  diciembreSaldo = 'diciembreSaldo',
  actualDebito = 'actualDebito',
  actualCredito = 'actualCredito',
  actualSaldo = 'actualSaldo',
}